requirements.txt
streamlit
openai
python-docx
docx2pdf
PyPDF2
pdfplumber
reportlab
stripe
requests
pandas
